test(bibleui): split external document import tests
Why:
- Continue the app-host test migration by moving external document import coverage into package-owned test targets.
- Keep routing and presentation assertions with BibleUI while placing TTF font filesystem contracts with SwordKit.

What Changed:
- Moved external document import routing and feedback tests to BibleUITests.
- Added TtfFontRepositoryTests under SwordKitTests for Android-style font addon installation and error contracts.
- Removed the old app-host test membership from the Xcode project and updated the migration plan.

Validation:
- xcodebuild test -scheme BibleUITests -only-testing:BibleUITests/ExternalDocumentImportTests passed.
- xcodebuild test -scheme SwordKitTests -only-testing:SwordKitTests/TtfFontRepositoryTests passed.
- xcodebuild test -scheme BibleUITests passed.
- xcodebuild test -scheme SwordKitTests passed.
- xcodebuild build-for-testing -scheme AndBible passed.
- swift package describe --type json passed.
- python3 scripts/check_repo_standards.py docblocks --all-files passed.
- python3 -m unittest discover -s scripts -p test_*.py passed.
- plutil -lint AndBible.xcodeproj/project.pbxproj passed.
- git diff --cached --check passed.
- GitNexus staged detect_changes reported low risk with no affected execution flows.
- Adversarial review completed; the stale snapshot docblock finding was fixed.

Impact:
- Removes another app-host test extension while preserving Android document import and font addon parity coverage in package targets.
- No production code changes.
